# Add indexes for vm_stats