--- a/engine/schema/src/main/java/com/cloud/upgrade/dao/Upgrade41900to41910.java
+++ b/engine/schema/src/main/java/com/cloud/upgrade/dao/Upgrade41900to41910.java
@@ -70,7 +70,7 @@ public class Upgrade41900to41910 implements DbUpgrade, DbUpgradeSystemVmTemplate
     }
 
     private void addIndexes(Connection conn) {
-        DbUpgradeUtils.addIndexIfNeeded(conn, "vm_stats", "vm_id", "timestamp");
+        DbUpgradeUtils.addIndexIfNeeded(conn, "vm_stats", "vm_id");
     }
 
     @Override